refactor: move validateProps function to utils
