+++ src/CircleLayout.tsx
@@ -1,6 +1,7 @@
 import React, { useMemo } from 'react';
 import { Animated, View } from 'react-native';
 
+import CircleLayoutArray from './CircleLayoutArray';
 import { CircleLayoutContext } from './CircleLayoutContext';
 import type {
   CircleLayoutContextType,
@@ -8,44 +9,7 @@ import type {
   CircleLayoutRef,
   Layout,
 } from './types';
-import CircleLayoutArray from './CircleLayoutArray';
-
-/**
- * Validates the props passed to the CircleLayout component and
- * returns the validated props. If any of the props are invalid, an error is thrown.
- * @param props The properties passed to the component
- * @returns The validated properties
- */
-export function validateProps(props: CircleLayoutProps): CircleLayoutProps & {
-  sweepAngle: number;
-  startAngle: number;
-} {
-  const errors: string[] = [];
-  if (props.components.length < 2) {
-    errors.push('At least two components need to be passed to CircleLayout');
-  }
-  if (props.radius <= 0) {
-    errors.push('Radius needs to be greater than 0');
-  }
-
-  if (props.sweepAngle === 0) {
-    errors.push('Sweep angle cannot be 0');
-  }
-
-  if (errors.length > 0) {
-    throw new Error(errors.join('\n'));
-  }
-
-  return {
-    ...props,
-    sweepAngle: props.sweepAngle
-      ? props.sweepAngle % (2 * Math.PI) === 0
-        ? 2 * Math.PI
-        : props.sweepAngle % (2 * Math.PI)
-      : 2 * Math.PI,
-    startAngle: (props.startAngle ?? 0) % (2 * Math.PI),
-  };
-}
+import { validateProps } from './utils/validation';
 
 /**
  * A component that places a list of components in a circular layout.
