+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export { interpolationWithFunction } from './animation';
 export { pointOnCircle } from './circle';
 export type { PointOnCircle } from './circle';
+export { validateProps } from './validation';
